Split one component (app.js) into multiple
