+++ src/components/Header.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Header = () => {
+  return (
+    <div>Header In here</div>
+  )
+}
+
+export default Header
